requirements.txt
numpy>=1.24
Pillow>=10.0
scipy>=1.11
requests>=2.31
